Move channel builder test utility into base and make it flexible for different models.
